fix: fix afs read panic
